Optional argument to match windows with given class by instance instead.
If the argument -i or --instance is provided then any window whose
*class* matches that argument will be treated as being identified by its
instance rather than its class or name.

This is particularly useful for browser-based web apps which may all
have the same class (that of the browser) but unique instance values.

This argument can be used multiple times to specify multiple strings to
match against window class.